:wrench: #304 テンプレート選択をrouter.jsでできるように改修。
