+++ resources/js/pages/baseshop/BaseApiIntegration.vue
@@ -1,0 +1,6 @@
+<template>
+  <div class="login-page">
+    <h1>ログイン</h1>
+    <!-- フォーム -->
+  </div>
+</template>
